SOP-Regulation-Verifier/requirements.txt
fastapi>=0.95.2
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6
markdown>=3.4.3
langchain-anthropic>=0.1.3
numpy>=1.25.0
scikit-learn>=1.4.2
faiss-cpu>=1.7.4
PyPDF2>=3.0.0
python-docx>=0.8.11
